+++ backend/src/routes/listingsRoutes.js
@@ -5,6 +5,7 @@ const wrapAsync = require("../utils/wrapAsync");
 const ExpressError = require("../utils/ExpressError");
 const router = express.Router();
 const {listingSchema }= require("../../schema");
+const isLoggedIn = require("../middleware/auth");
 
 
 
@@ -40,8 +41,14 @@ router.get("/:id", wrapAsync(async (req, res) => {
         throw new ExpressError(400, "Invalid ID");
     }
 
-    const listingdata = await Listing.findById(id).populate("reviews");
-
+    const listingdata = await Listing.findById(id) .populate("owner", "username name")   
+        .populate({
+            path: "reviews",
+            populate: {
+                path: "user",
+                select: "username name"  
+            }
+        }); 
     if (!listingdata) {
         throw new ExpressError(404, "Listing not found");
     }
@@ -51,6 +58,7 @@ router.get("/:id", wrapAsync(async (req, res) => {
 
 
 router.post("/add", 
+     isLoggedIn,
    validateListing ,
     wrapAsync(async (req, res) => {
 
@@ -58,20 +66,29 @@ router.post("/add",
    
    
     const newlisting = new Listing(req.body.listing);
+     newlisting.owner = req.user._id;
 
     const savedListing = await newlisting.save();
     res.status(201).json(savedListing);
 
 
-
 }));
 router.put("/:id", 
+    isLoggedIn,
     validateListing ,
      wrapAsync(async (req, res) => {
 
     const { id } = req.params;
     
+  const listing = await Listing.findById(id);
 
+    if (!listing) {
+      throw new ExpressError(404, "Listing not found");
+    }
+
+    if (!listing.owner.equals(req.user._id)) {
+      return res.status(403).json({ message: "You are not allowed to edit this listing" });
+    }
 
     const uplistings = await Listing.findByIdAndUpdate(id, { ...req.body.listing });
     if (!uplistings) {
@@ -82,9 +99,22 @@ router.put("/:id",
 }));
 
 
-router.delete("/:id", wrapAsync(async (req, res) => {
+router.delete("/:id",
+    isLoggedIn,
+     wrapAsync(async (req, res) => {
 
     const { id } = req.params;
+
+    const listing = await Listing.findById(id);
+
+    if (!listing) {
+      throw new ExpressError(404, "Listing not found");
+    }
+    if (!listing.owner.equals(req.user._id)) {
+      return res.status(403).json({
+        message: "You are not allowed to delete this listing"
+      });
+    }
     let dlisting = await Listing.findByIdAndDelete(id);
 
     if (!dlisting) {
@@ -97,8 +127,21 @@ router.delete("/:id", wrapAsync(async (req, res) => {
 
 
 
+router.get("/mylisting/:id", 
+    isLoggedIn,
+    wrapAsync(async (req, res) => {
+    const { id } = req.params;
+    console.log(req.params);
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        throw new ExpressError(400, "Invalid ID");
+    }
+    const listings = await Listing.find({ owner: id });
+    if (!listings || listings.length === 0) {
+        throw new ExpressError(404, "No listings found for this user");
+    }
+    res.status(200).json(listings);
+}));
 
 
-
-
-module.exports = router;
+module.exports = router;
